refactor : 파일 구조 개선
