--- src/app.module.ts
@@ -1,33 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { ConfigModule } from '@nestjs/config';
-import { UserModule } from './user/user.module';
-import { AuthModule } from './auth/auth.module';
-import { EventModule } from './events/event.module';
-import { ChatModule } from './chat/chat.module';
-import { FriendsModule } from './friends/friends.module';
-
-@Module({
-  imports: [
-    // ConfigModule로 환경 변수 로드
-    ConfigModule.forRoot({
-      isGlobal: true, // 모든 모듈에서 환경 변수를 사용할 수 있게 설정
-    }),
-    TypeOrmModule.forRoot({
-      type: 'postgres',
-      host: process.env.DB_HOST,
-      port: process.env.DB_PORT ? +process.env.DB_PORT : 5432,
-      username: process.env.DB_USERNAME,
-      password: process.env.DB_PASSWORD,
-      database: process.env.DB_DATABASE,
-      entities: [__dirname + '/**/*.entity{.ts,.js}'],
-      synchronize: true,
-    }),
-    EventModule,
-    UserModule,
-    AuthModule,
-    ChatModule,
-    FriendsModule,
-  ],
-})
-export class AppModule {}
