+++ src/module/auth/auth.interface.ts
