+++ src/module/chat/entities/chat-room.entity.ts
