+++ src/module/events/entities/event.entity.ts
