+++ src/module/friends/dto/request/find-friend-by-email.dto.ts
