+++ src/module/friends/entities/friends.entity.ts
@@ -1,4 +1,4 @@
-import { User } from 'src/user/user.entity';
+import { User } from 'src/module/user/user.entity';
 import { Entity, PrimaryGeneratedColumn, ManyToOne, JoinColumn } from 'typeorm';
 
 @Entity()
